requirements.txt
asgiref==3.5.2
camelcase==0.2
click==8.1.3
colorama==0.4.6
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dotenv==0.0.5
Flask==2.2.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
PyJWT==1.7.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
Werkzeug==2.2.2
